server/requirements.txt
asgiref==3.2.10
astroid==2.3.2
autopep8==1.4.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.1
cryptography==3.1.1
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.2.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mpmath==1.1.0
oauthlib==3.1.0
pbr==5.4.3
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.4.1
stevedore==1.31.0
sympy==1.4
urllib3==1.25.11
virtualenv==16.7.6
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
